Refactor CLI commands into a lib for integration testing and publishing command-level functions.
